feat: Add support for setting max pods per node [DEVOP-7311]
To improve operational density, and due to current issues being
experienced in our `dev` environment which is only running 3 nodes,
we want to be able to increase the default cluster value for the max
number of pods running on a single node.

This will allow us to shift from our current `110` pod limit, up to the
hard cap of `256` pods per node. This does increase the CIDR size
required by the node from a `/24` to a `/23`, but each environment
operates with a `/16` block, so there is plenty of space to be had.

Refs: #DEVOP-7311

Signed-off-by: Christian Witts <christian@honestbank.com>